Refactor maxFeaturesInView parsing to use Number for consistency and move rectange utilities to utils
